Split ui.js out
