+++ cart-pole/package.json
@@ -9,7 +9,7 @@
     "node": ">=8.9.0"
   },
   "dependencies": {
-    "@tensorflow/tfjs": "^0.11.7",
+    "@tensorflow/tfjs": "^0.12.0",
     "vega-embed": "^3.0.0"
   },
   "scripts": {
@@ -32,6 +32,7 @@
     "karma-chrome-launcher": "~2.2.0",
     "karma-firefox-launcher": "~1.1.0",
     "karma-jasmine": "~1.1.1",
-    "parcel-bundler": "~1.8.1"
+    "parcel-bundler": "~1.8.1",
+    "typescript": "^2.9.2"
   }
 }
